+++ App.js
@@ -4,42 +4,11 @@ import { NavigationContainer } from '@react-navigation/native'
 import Tabs from './src/components/Tabs'
 import * as Location from 'expo-location'
 import { WEATHER_API_KEY } from '@env'
+import { useGetWeather } from './src/hooks/useGetWeather'
 
 const App = () => {
-  const [isLoading, setIsLoading] = useState(true)
-  const [error, setError] = useState(null)
-  const [weather, setWeather] = useState([])
-  const [lat, setLat] = useState([])
-  const [lon, setLon] = useState([])
-  const fetchWeatherData = async () => {
-    try {
-      const res = await fetch(`api.openweathermap.org/data/2.5/forecast?lat=${lat}&lon=${lon}&appid=${WEATHER_API_KEY}`)
-      const data = await res.json()
-      setWeather(data)
-    } catch (e) {
-      setError('could not fetch weather data')
-    } finally {
-      setIsLoading(false)
-  }}
-  useEffect(() => {
-    ;(async () => {
-      let { status } = await Location.requestForegroundPermissionsAsync()
-      if (status !== 'granted') {
-        setError('permission to access location was denied')
-        return
-      }
-      let location = await Location.getCurrentPositionAsync({})
-      setLat(location.coords.latitude)
-      setLon(location.coords.longitude)
-      await fetchWeatherData()
-    })()
-  }, [lat, lon])
-
-  if (weather.length) {
-    console.log('weather: ', weather)
-  } else {
-    console.log('no weather')
-  }
+  const [isLoading, error, weather] = useGetWeather()
+  console.log(isLoading, error, weather)
 
   if (isLoading) {
     return (
